[issue033] cplusplus-basics-language: Supress spurious cross-compile warning
FEATURE(s):
[issue033] cplusplus-basics-language: Explore the C++ programming
  language.

FILE(s)
[modified]TODO.txt: Add issue033
[new]cplusplus/basics/.idea/.gitignore:
  Empty gitignore file for the sake of consistency
[modified]cplusplus/basics/Makefile
  Add -fno-poison-system-directories to stop clang
  from complaining about inclusion of system directories.
  That warning is only relevent cross-compiled code.
[modified]cplusplus/basics/basics/strings.cc
  Remove pragmas for no-poison-system-drectories
  They don't seemd to be working when specified as #pragma
  Need to investigate why.

DESCRIPTION
Supress spurious cross-compile warning

Because I've enabled all possible warnings to catch any obscure
errors, the compiler will warn about things that are not problems.
One of them is complaining about including system headers
in a cross-compilation situation. This is a spurious warning,
as we are not actually cross-compiling the code.

So we supress this warning the Makefile. The previous
attempt to use #pragma to do the supression did not
work for reasons that I do not yet understand.