[issue031] javascript-basics-language: Add IntelliJ IDEA inspection profile
FEATURE(s)
[issue031] javascript-basics-language: Explore the basics
  of the javascript language..

FILE(s)
[modified]javascript/basics/00-setup/README.md
  Improve instructions on how to open the IntelliJ IDEA
  project.

DESCRIPTION
Add IntelliJ IDEA inspection profile
For some reason IntelliJ IDEA keeps insisting that
inspection profiles for the python project need to
be added to the javascript-basics-language project.
Not sure why. But adding it so as to stop it from
showing up in commit lists all the time. It should
be harmless.