issue130-python-parse-dralithus-cmdline: Remove unnecessary code in the catch block of parse_command_line()
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/parse_command_line.py:
  Remove unnecessary code in the catch block of parse_command_line().

DESCRIPTION
Remove unnecessary code in the catch block of parse_command_line().

The code that sets verbosity, environment and parameters is not
necessary, since they are already set when the options dictionary and
parameters list are initialized at the top of the function. The only
thing that needs to be changed is the help value: from True to False.