issue130-python-parse-dralithus-cmdline: Modify add_to_options_dict() to only accept integer values for the verbosity option
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/parse_command_line.py:
  Modify add_to_options_dict() to only accept integer values for the
  verbosity option

DESCRIPTION
Modify add_to_options_dict() to only accept integer values for the verbosity option.

Raise an exception if the value of the verbosity option is not an integer.
Otherwise, just add the value of the verbosity option to the value in
the options dictionary if it is an integer.