issue130-python-parse-dralithus-cmdline: Fix big in is_short option()
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/parse_command_line.py:
  Fix bug in is_short option()

DESCRIPTION
Fix bug in is_short option().

The test should be for isalnum not (isalpha() or isdigit()).
The former will correctly verify that the string is alphanumeric,
while the latter will only check if the string is either all alphabetic
or all numeric.