issue130-python-parse-dralithus-cmdline: Refactor code to move test case generation functions to module level
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/test/test_parse_command_line.py:
Refactor code to move test case generation functions to module level
instead of static methods in the class

DESCRIPTION
Refactor code to move test case generation functions to module level
instead of static methods in the class.

It seems like something in the parameterized
framework is interfering with the ability of a static method in the
class to call another static method.