issue130-python-parse-dralithus-cmdline: Modify test_debug so that it's decorator causes it to be be executed only if the module variable DEBUG_TEST_NUMBER is set to an integer
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/test/test_parse_command_line.py:
  Modify test_debug so that it's decorator causes it to be be executed
  only if the module variable DEBUG_TEST_NUMBER is set to an integer

DESCRIPTION
Modify test_debug so that it's decorator causes it to be be executed
only if the module variable DEBUG_TEST_NUMBER is set to an integer.

This is an improvement to the test_debug function in that one
does not need to modify the code itself, just a global variable.