issue130-python-parse-dralithus-cmdline: Modify test_debug() to get the DEBUG_TEST_NUMBER flag from an environment variable instead of a hardcoded value
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/test/test_parse_command_line.py:
  Modify test_debug() to get the DEBUG_TEST_NUMBER flag from an
  environment variable instead of a hardcoded value

DESCRIPTION
Modify test_debug() to get the DEBUG_TEST_NUMBER flag from an
environment variable instead of a hardcoded value.

This is even better improvement over the previous commit, as it
allows for the debug test to be run without any modification to
the code. The test will now use the value of the environment
variable DEBUG_TEST_NUMBER, if it is set, otherwise the test
will be skipped.