issue130-python-parse-dralithus-cmdline: Refactor test_parse_command_line.py to use list[TestCaseData] everywhere except for all_cases()
FEATURE(s)
issue130-python-parse-dralithus-cmdline: Experimental parser for the dralithus command line

FILE(s)
[modified]python/parse-dralithus-cmdline/test/test_parse_command_line.py:
  Refactor test_parse_command_line.py to use list[TestCaseData]
  everywhere except for all_cases()

DESCRIPTION
Refactor test_parse_command_line.py to use list[TestCaseData]
everywhere except for all_cases().

all_cases() is an exception, because the @parameterized.expand() decorator
expects a list of tuples where the first element is the name of the
test case, and the second element is the case itself.