issue146-python-parsec: Add commented out raise statements for Option.supported_short_flags() and Option.supported_long_flags()
FEATURE(s)
issue146-python-parsec: Experimental standalone command line parser library

FILE(s)

DESCRIPTION
Add commented out raise statements for Option.supported_short_flags()
and Option.supported_long_flags().

The classmethod in the base class has confusing semantics.
In the base classs it is implemented as a function that returns
all the flags that are supported by every sub-class of Options.
In sub-classes this class method is overridden to return just
the flags supported by the option.

It's better to separate that functionality and keep the two
functions purely abstract in the base class.