issue146-python-parsec: Refactor parse() into class Parser.
FEATURE(s)
issue146-python-parsec: Experimental standalone command line parser library

FILE(s)
[modified]python/parsec/src/parsec/command_line.py:
  Refactor parse() into class Parser.
[modified]python/parsec/tests/test_parsec/test_command_line.py:
  Create class TestParser to test class Parser.
  Move the test_parse method from TestCommandLine into TestParser
  Modify the call toe execute to take a lamdba that wraps the
  creation of a parser object and the call to parse into single
  function

DESCRIPTION
Refactor parse() into class Parser.

This allows encapsulation of the helper functions as private methods,
and will eventually allow the parser object to store state of the
parser in an object of the class. Furthermore in future, one
could enable the parser to be initialized with a collection of
supported options, thus making it possible to construct a custom
command line parser for different CLI programs.