issue146-python-parsec: Move _make_verbosity back to being a module level function
FEATURE(s)
issue146-python-parsec: Experimental standalone command line parser library

FILE(s)
[modified]python/parsec/src/parsec/command_line.py:
  Move _make_verbosity back to being a module level function

DESCRIPTION
Move _make_verbosity back to being a module level function

It is used in both class CommandLine for the verbosity property
and in class Parser. So for now just make it a module function.
In future, it would make sense for it to be a method in Parser
and passed to the constructor of the CommandLine class.