issue149-python-packages.sh: Use an env trampoline to use the default bash
FEATURE(s)
issue149-python-packages.sh: Use an env trampoline to use the default bash

FILE(s)
[modified]TODO.txt: Add issue 149 as DONE
[modified]python/bin/packages.sh
Use an env trampoline to use the default bash

DESCRIPTION
Use an env trampoline to use the default bash.

On macOS the defaul /bin/bash is very ancient and does not
support options like -v. To overcome this we want to take
the bash version specified in the user's environment. On
my Mac this is the the bash version provided by macports.
Which is located and /opt/local/bin/bash, and is available
on the path.

This is not a universal solution to finding the right bash,
but it will do for now.